refactor(engine): :zap: using types from entities package
