+++ packages/engine/src/helpers/computeExperienceMetadata.test.ts
@@ -10,7 +10,7 @@ import {
 } from "../mocks/sample-experience-session"
 import computeExperienceMetadata from "./computeExperienceMetadata"
 import JSONConfig from "../classes/JSONConfig"
-import { IFireboltSession, IExperienceMetadata } from "../interfaces/IEngine"
+import { IFireboltSession, IExperienceMetadata } from "../types"
 
 interface ISessionTestCase {
   label: string
