+++ packages/engine/src/helpers/getIsFieldsValidationNeeded.ts
@@ -1,7 +1,7 @@
 import {
   IFireboltSession,
   IExperienceProceedPayload,
-} from "../interfaces/IEngine"
+} from "../types"
 import { equals } from "ramda"
 
 interface IIsFieldsValidationNeeded {
