+++ packages/engine/src/helpers/getReturningStepFilled.test.ts
@@ -1,6 +1,6 @@
 import getReturningStepFilled from "./getReturningStepFilled"
 import sampleExperience from "../mocks/sample-experience"
-import { IFireboltSession } from "../interfaces/IEngine"
+import { IFireboltSession } from "../types"
 
 describe("test autofill work", () => {
   test("Function should return a new step filled", () => {
