+++ packages/engine/src/helpers/getShouldUseDecisionHandler.ts
@@ -1,5 +1,5 @@
 import EngineError from "../classes/EngineError"
-import { IDecisionHandlerConfig } from "../types"
+import { IDecisionHandlerConfig } from "@iq-firebolt/entities"
 
 export default function shouldUseDecisionHandler(
   decisionHandlerConfig: IDecisionHandlerConfig | undefined,
