+++ packages/engine/src/mocks/storage-session-3-steps.js
@@ -4,7 +4,7 @@
  */
 
 /**
- * @typedef {import('../interfaces/IEngine').IFireboltSession} IFireboltSession
+ * @typedef {import('../types').IFireboltSession} IFireboltSession
  */
 
 /**
